+++ src/dallassensor.h
@@ -91,11 +91,11 @@ class DallasSensor {
         return sensors_;
     }
 
-    uint16_t reads() {
+    uint32_t reads() {
         return sensorreads_;
     }
 
-    uint16_t fails() {
+    uint32_t fails() {
         return sensorfails_;
     }
 
@@ -171,8 +171,8 @@ class DallasSensor {
     uint8_t  dallas_gpio_ = 0;
     bool     parasite_    = false;
     bool     changed_     = false;
-    uint16_t sensorfails_ = 0;
-    uint16_t sensorreads_ = 0;
+    uint32_t sensorfails_ = 0;
+    uint32_t sensorreads_ = 0;
     int8_t   scanretry_   = 0;
 };
